requirements.txt
aiosqlite==0.20.0        # 비동기 SQLite 제어
beautifulsoup4==4.12.3   # HTML 파싱
aiohttp==3.9.5           # 비동기 HTTP 요청 (크롤링용)
aiogram==3.4.1           # 비동기 텔레그램 봇 API 프레임워크
google-generativeai==0.5.2 # 구글 Gemini LLM API
python-dotenv==1.0.1     # 환경변수 로드